# openapi/eth402.yaml
openapi: 3.1.0
info:
  title: ETH402 service API
  version: 0.6.0
  description: >-
    ETH402 merchant and operational APIs plus the standard x402 v2 facilitator
    verification and settlement surface. Merchant authentication is separate
    from x402 payment authorization. Settlement requires a prior verification
    and a configured settlement signer.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://eth402.org
paths:
...
  /verify:
    post:
      operationId: x402Verify
      summary: Verify an x402 v2 EIP-3009 authorization without broadcasting
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/X402VerifyRequest" }
      responses:
        "200":
          description: Protocol verification result; invalid authorizations also use this status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402VerifyResponse" }
        "400":
          description: The outer request is not valid x402 v2 JSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402VerifyResponse" }
        "503":
          description: Verification could not safely complete due to a dependency failure.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402VerifyResponse" }
  /settle:
    post:
      operationId: x402Settle
      summary: Settle a previously verified x402 v2 EIP-3009 payment
      description: >-
        Requires a prior successful /verify for the same payment: admission is
        gated on the durable verification record, which is what binds the
        recipient to an active registered merchant (ADR-0004). The request
        shape is identical to /verify and is parsed by the same strict rules.
        On success the facilitator has broadcast transferWithAuthorization and
        returns its transaction hash; confirmation completes asynchronously
        (12 confirmations by default). Calls are idempotent per payment: a
        repeat settles to the same transaction hash. Policy rejections return
        HTTP 200 with success=false and a stable errorReason.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/X402SettleRequest" }
      responses:
        "200":
          description: Settlement outcome, including policy rejections with success=false.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402SettleResponse" }
        "400":
          description: The outer request is not valid x402 v2 JSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402SettleResponse" }
        "503":
          description: Settlement is unavailable (signer disabled or a dependency failure).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/X402SettleResponse" }
  /v1/merchants/register:
    post:
      operationId: registerMerchant
      summary: Start merchant registration
      description: Returns the same accepted response for new and duplicate valid registrations.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegistrationRequest" }
      responses:
        "202":
          description: Email verification is pending.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pending" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/merchants/verify-email:
    post:
      operationId: verifyMerchantEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, minLength: 20 }
      responses:
        "200":
          description: Email token was consumed.
          content:
            application/json:
              schema:
                type: object
                required: [merchant_id, status]
                properties:
                  merchant_id: { type: string, format: uuid }
                  status: { const: email_verified }
        "400": { $ref: "#/components/responses/Error" }
  /v1/merchants/wallet-challenge:
    post:
      operationId: createInitialWalletChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [merchant_id]
              properties:
                merchant_id: { type: string, format: uuid }
      responses:
        "201":
          description: EIP-4361 challenge for the registered recipient.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WalletChallenge" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/merchants/verify-wallet:
    post:
      operationId: verifyInitialWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletVerificationRequest" }
      responses:
        "200":
          description: Merchant activated; the API key is returned exactly once.
          content:
            application/json:
              schema:
                type: object
                required: [status, api_key]
                properties:
                  status: { const: active }
                  api_key: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/me:
    get:
      operationId: getMerchant
      security: [{ MerchantAPIKey: [] }]
      responses:
        "200":
          description: Authenticated merchant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Merchant" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/api-keys:
    get:
      operationId: listAPIKeys
      security: [{ MerchantAPIKey: [] }]
      responses:
        "200":
          description: Key metadata; full secret values are never listed.
          content:
            application/json:
              schema:
                type: object
                required: [keys]
                properties:
                  keys:
                    type: array
                    items: { $ref: "#/components/schemas/APIKey" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
    post:
      operationId: createAPIKey
      security: [{ MerchantAPIKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "201":
          description: New full API key returned exactly once.
          content:
            application/json:
              schema:
                type: object
                required: [key, api_key]
                properties:
                  key: { $ref: "#/components/schemas/APIKey" }
                  api_key: { type: string }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/api-keys/{id}:
    delete:
      operationId: revokeAPIKey
      security: [{ MerchantAPIKey: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Key revoked. }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/me/recipient-change:
    post:
      operationId: createRecipientChangeChallenge
      security: [{ MerchantAPIKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [new_address]
              properties:
                new_address: { $ref: "#/components/schemas/EthereumAddress" }
      responses:
        "201":
          description: Challenge bound to the new recipient address.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WalletChallenge" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/me/recipient-change/verify:
    post:
      operationId: verifyRecipientChange
      security: [{ MerchantAPIKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [challenge_id, message, signature]
              properties:
                challenge_id: { type: string, format: uuid }
                message: { type: string }
                signature: { type: string }
      responses:
        "200":
          description: Recipient changed and history appended.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { const: recipient_changed }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /v1/admin/merchants/{id}/suspend:
    post:
      operationId: suspendMerchant
      description: Operator-only endpoint. Disabled when no operator token is configured.
      security: [{ OperatorToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason_code]
              properties:
                reason_code: { type: string, minLength: 1 }
      responses:
        "200":
          description: Merchant suspended.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { const: suspended }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/admin/merchants/{id}/reinstate:
    post:
      operationId: reinstateMerchant
      description: Operator-only endpoint. Disabled when no operator token is configured.
      security: [{ OperatorToken: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Merchant reinstated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { const: active }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /stats:
    get:
      operationId: publicStats
      summary: Aggregate public facilitator statistics
      description: >-
        Cached aggregate data. Only confirmed settlements contribute volume.
        Volume figures are omitted unless the operator sets
        ETH402_PUBLISH_STATS_VOLUME: a cumulative total polled over time yields its
        own deltas, and a delta spanning a single settlement is that payment's exact
        amount. Health is probed on the cache schedule, not per request, so this
        public endpoint cannot be used to drive database or RPC load.
      responses:
        "200":
          description: Version 2 statistics.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsV2"
        "429":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /status:
    get:
      operationId: statusPage
      summary: Human-readable status page
      description: >-
        Renders the same cached snapshot as /stats as a self-contained HTML page —
        no external stylesheet, script, font, or image, so it keeps working during
        the network failures it exists to report. It cannot disclose more than
        /stats does.
      responses:
        "200":
          description: Status page.
          content:
            text/html:
              schema: { type: string }
        "503":
          description: Aggregates could not be read; the page reports an outage.
          content:
            text/html:
              schema: { type: string }
...
components:
  securitySchemes:
    MerchantAPIKey:
      type: http
      scheme: bearer
      bearerFormat: ETH402 API key
    OperatorToken:
      type: apiKey
      in: header
      name: X-Operator-Token
  responses:
    Error:
      description: Structured service error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    X402VerifyRequest:
      type: object
      additionalProperties: false
      required: [x402Version, paymentPayload, paymentRequirements]
      properties:
        x402Version: { const: 2 }
        paymentPayload: { $ref: "#/components/schemas/X402PaymentPayload" }
        paymentRequirements: { $ref: "#/components/schemas/X402PaymentRequirements" }
    X402VerifyResponse:
      type: object
      required: [isValid]
      properties:
        isValid: { type: boolean }
        invalidReason: { type: string }
        invalidMessage: { type: string }
        payer: { $ref: "#/components/schemas/EthereumAddress" }
        extensions: { type: object }
        extra: { type: object }
    X402SettleRequest:
      type: object
      additionalProperties: false
      required: [x402Version, paymentPayload, paymentRequirements]
      properties:
        x402Version: { const: 2 }
        paymentPayload: { $ref: "#/components/schemas/X402PaymentPayload" }
        paymentRequirements: { $ref: "#/components/schemas/X402PaymentRequirements" }
    X402SettleResponse:
      type: object
      required: [success, transaction, network]
      properties:
        success: { type: boolean }
        errorReason:
          type: string
          enum:
            - invalid_request
            - payment_not_found
            - payment_not_verified
            - recipient_not_registered
            - authorization_expiring
            - merchant_quota_exceeded
            - facilitator_quota_exceeded
            - simulation_reverted
            - broadcast_failed
            - settlement_unavailable
        errorMessage: { type: string }
        payer: { $ref: "#/components/schemas/EthereumAddress" }
        transaction:
          type: string
          description: Broadcast transaction hash; empty when success is false.
          pattern: "^(0x[0-9a-f]{64})?$"
        network: { const: eip155:1 }
        amount: { type: string, pattern: "^[0-9]+$" }
        extensions: { type: object }
        extra: { type: object }
    EthereumAddress:
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
    RegistrationRequest:
      type: object
      additionalProperties: false
      required: [name, business_email, recipient_address, accept_terms]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        business_email: { type: string, format: email }
        recipient_address: { $ref: "#/components/schemas/EthereumAddress" }
        accept_terms: { const: true }
        website: { type: string, format: uri }
        description: { type: string }
    Pending:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { const: verification_pending }
    WalletChallenge:
      type: object
      additionalProperties: false
      required: [id, message, address, action, expires_at]
      properties:
        id: { type: string, format: uuid }
        message: { type: string }
        address: { $ref: "#/components/schemas/EthereumAddress" }
        action: { enum: [verify-recipient, change-recipient] }
        expires_at: { type: string, format: date-time }
    WalletVerificationRequest:
      type: object
      additionalProperties: false
      required: [merchant_id, challenge_id, message, signature]
      properties:
        merchant_id: { type: string, format: uuid }
        challenge_id: { type: string, format: uuid }
        message: { type: string }
        signature: { type: string }
    Merchant:
      type: object
      additionalProperties: false
      required: [id, name, business_email, recipient_address, status, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        business_email: { type: string, format: email }
        recipient_address: { $ref: "#/components/schemas/EthereumAddress" }
        status: { enum: [pending, manual_review, active, suspended, rejected] }
        website: { type: string, format: uri }
        description: { type: string }
        email_verified_at: { type: string, format: date-time }
        wallet_verified_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    APIKey:
      type: object
      additionalProperties: false
      required: [id, name, prefix, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }